Create BorderedButton class
Instead of setting the layer properties in the
DCMShowDetailViewController, make a custom UIButton subclass that
knows how to set the properties on itself. Bonus: by setting the
border color in `tintColorDidChange`, the border will also desaturate
along with the button text when the tint color changes because, for
example, a modal alert is displayed.